+++ CordialSDK/CordialSDK/Endpoints/UpsertContactCart/Helper/CartItem.swift
@@ -22,7 +22,7 @@ import Foundation
     var timestamp: String
     let attr: Dictionary<String, String>?
     let images: [String]?
-    let properties: Dictionary<String, String>?
+    let properties: Dictionary<String, Any>?
         
     enum Key: String {
         case productID = "productID"
@@ -40,11 +40,11 @@ import Foundation
         case properties = "properties"
     }
     
-    @objc public convenience init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qtyNumber: NSNumber?, itemPriceNumber: NSNumber?, salePriceNumber: NSNumber?, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, String>?) {
+    @objc public convenience init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qtyNumber: NSNumber?, itemPriceNumber: NSNumber?, salePriceNumber: NSNumber?, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, Any>?) {
         self.init(productID: productID, name: name, sku: sku, category: category, url: url, itemDescription: itemDescription, qty: qtyNumber?.int64Value, itemPrice: itemPriceNumber?.doubleValue, salePrice: salePriceNumber?.doubleValue, attr: attr, images: images, properties: properties)
     }
     
-    public init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qty: Int64?, itemPrice: Double?, salePrice: Double?, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, String>?) {
+    public init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qty: Int64?, itemPrice: Double?, salePrice: Double?, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, Any>?) {
         self.productID = productID
         self.name = name
         self.sku = sku
@@ -86,7 +86,7 @@ import Foundation
         aCoder.encode(self.properties, forKey: Key.properties.rawValue)
     }
     
-    private init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qty: Int64?, itemPrice: Double?, salePrice: Double?, timestamp: String, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, String>?) {
+    private init(productID: String, name: String, sku: String, category: String?, url: String?, itemDescription: String?, qty: Int64?, itemPrice: Double?, salePrice: Double?, timestamp: String, attr: Dictionary<String, String>?, images: [String]?, properties: Dictionary<String, Any>?) {
         self.productID = productID
         self.name = name
         self.sku = sku
@@ -115,7 +115,7 @@ import Foundation
         let timestamp = aDecoder.decodeObject(forKey: Key.timestamp.rawValue) as! String
         let attr = aDecoder.decodeObject(forKey: Key.attr.rawValue) as! Dictionary<String, String>?
         let images = aDecoder.decodeObject(forKey: Key.images.rawValue) as! [String]?
-        let properties = aDecoder.decodeObject(forKey: Key.properties.rawValue) as! Dictionary<String, String>?
+        let properties = aDecoder.decodeObject(forKey: Key.properties.rawValue) as! Dictionary<String, Any>?
         
         self.init(productID: productID, name: name, sku: sku, category: category, url: url, itemDescription: itemDescription, qty: qty, itemPrice: itemPrice, salePrice: salePrice, timestamp: timestamp, attr: attr, images: images, properties: properties)
     }